fast_api/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
passlib[bcrypt]
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.0
h11==0.14.0
idna==3.7
jwt==1.3.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-jose==3.3.0
requests==2.31.0
router==0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1